[Fix][TIR] LowerThreadAllreduce with correct thread mask
This PR fixes a bug in the LowerThreadAllreduce pass.

Prior to this PR, in multi-group settings, the thread mask is not
correctly set: when the reduction extent is 32, the thread mask will
always be 0. This bug was not spotted because even when the mask is 0,
the CUDA program still gives correct result. But in any way, having
the zero mask is dangerous and should be fixed.

With this PR, we will only update the mask with bitwise-and when the
reduction extent is less than the warp size.